% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{approval}
\alias{approval}
\title{U.S. Presidential Approval Data}
\format{
A table containing 8 variables ("month", "year", "approve",
"disapprove", "unsure", "sept.oct.2001", "iraq.war", and
"avg.price") and 65 observations.
}
\source{
ICPSR
}
\description{
U.S. Presidential Approval Data
}
\keyword{datasets}
