% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avg}
\alias{avg}
\title{Compute central tendancy as approrpriate to data type}
\usage{
avg(val)
}
\arguments{
\item{val}{a vector of values}
}
\value{
a mean (if numeric) or a median (if ordered) or mode (otherwise)
}
\description{
Compute central tendancy as approrpriate to data type
}

