% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bivariate}
\alias{bivariate}
\title{Sample data for bivariate probit regression}
\format{
A table containing 6 variables ("y1", "y2", "x1", "x2", "x3", and "x4") and 78 observations.
}
\source{
This is a cleaned and relabelled version of the sanction
data set, available in Zelig.
}
\description{
Sample data for bivariate probit regression
}
\references{
Martin, Lisa (1992).  \emph{Coercive Cooperation:
Explaining Multilateral Economic Sanctions}, Princeton: Princeton University Press.
}
\keyword{datasets}
