% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap model coefficients}
\usage{
bootstrap(obj, ...)
}
\arguments{
\item{obj}{a fitted model object fitted using \code{lm} and many using
\code{glm}. Note: more intended in future Zelig releases.}

\item{...}{other arguments passed to \code{zelig}.}
}
\description{
Bootstrap model coefficients
}
\examples{

set.seed(10)
lm.out <- bootstrap(lm(Fertility ~ Education, data = swiss))
summary(lm.out)

set.seed(10)
z.out <- zelig(Fertility ~ Education, data = swiss, model = "ls", bootstrap = TRUE)
summary(z.out)

all.equal(coef(lm.out), coef(z.out))

}
\author{
Christopher Gandrud and Ista Zahn
}
