% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ci_check}
\alias{ci_check}
\title{Convert \code{ci} interval from percent to proportion and check if valid}
\usage{
ci_check(x)
}
\arguments{
\item{x}{numeric. The central interval to return, expressed on the \code{(0, 100]}
or the equivalent \code{(0, 1]} interval.}
}
\description{
Convert \code{ci} interval from percent to proportion and check if valid
}
\keyword{internal}
