% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CigarettesSW}
\alias{CigarettesSW}
\title{Cigarette Consumption Panel Data}
\format{A data set with 96 observations and 9 variables}
\source{
From Christian Kleiber and Achim Zeileis (2008). Applied
Econometrics with R. New York: Springer-Verlag. ISBN 978-0-387-77316-2. URL
\url{https://CRAN.R-project.org/package=AER}
}
\description{
Cigarette Consumption Panel Data
}
\keyword{datasets}
