% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster.formula}
\alias{cluster.formula}
\title{Describe Here}
\usage{
cluster.formula(formula, cluster)
}
\arguments{
\item{formula}{Describe Here}

\item{cluster}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\keyword{internal}

