% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coalition2}
\alias{coalition2}
\title{Coalition Dissolution in Parliamentary Democracies, Modified Version}
\format{
A table containing 8 variables ("duration", "ciep12",
"invest", "fract", "polar", "numst2", "crisis", "country") and 314 
observations.  For variable descriptions, please refer to King,
Alt, Burns and Laver (1990).
}
\source{
ICPSR
}
\description{
This data set contains survival data on government
coalitions in parliamentary democracies (Belgium, Canada, Denmark, 
Finland, France, Iceland, Ireland, Israel, Italy, Netherlands,
Norway, Portugal, Spain, Sweden, and the United Kingdom) for the
period 1945-1987. Country indicator variables are included in the
sample data.
}
\keyword{datasets}
