\name{createJSON}
\alias{createJSON}
\title{Utility function for constructing JSON file that encodes the hierarchy of available statistical models in Zelig}
\usage{
createJSON()
}
\value{
Returns TRUE on successful completion of json file
}
\description{
Utility function for construction a JSON file that encodes the hierarchy of available statistical models.  
}
\author{
Christine Choirat, Vito D'Orazio
}

