% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{df.residual,Zelig-method}
\alias{df.residual,Zelig-method}
\title{Method for extracting residual degrees-of-freedom from Zelig objects}
\usage{
\S4method{df.residual}{Zelig}(object)
}
\arguments{
\item{object}{An Object of Class Zelig}
}
\description{
Method for extracting residual degrees-of-freedom from Zelig objects
}
