\name{eidat}

\alias{eidat}

\title{Simulation Data for Ecological Inference}

\description{
  This dataframe contains a simulated data set to illustrate the models
  for ecological inference.  
}

\usage{data(eidat)}

\format{
  A table containing 4 variables ("t0", "t1", "x0", "x1") and 10 
observations.
}

\keyword{datasets}
