% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_grid_setrange}
\alias{expand_grid_setrange}
\title{Convenience function for setrange and setrange}
\usage{
expand_grid_setrange(x)
}
\arguments{
\item{x}{data passed to setrange or setrange1}
}
\description{
Convenience function for setrange and setrange
}
\keyword{internal}
