% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{extract_setrange}
\alias{extract_setrange}
\title{Extract setrange to return as tidy formatted data frame}
\usage{
extract_setrange(obj, which_range = "range", only_setx = FALSE)
}
\arguments{
\item{obj}{a zelig object containing a range of simulated quantities of
interest}

\item{which_range}{character string either \code{'range'} or \code{'range1'}
indicating whether to extract the first or second set of fitted values}

\item{only_setx}{logical whether or not to only extract `setx`` values.}
}
\description{
Extract setrange to return as tidy formatted data frame
}
\seealso{
\code{\link{zelig_qi_to_df}}
}
\author{
Christopher Gandrud
}
\keyword{internal}
