% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{factor_coef_combine}
\alias{factor_coef_combine}
\title{Return individual factor coefficient fitted values to single factor variable}
\usage{
factor_coef_combine(obj, fitted)
}
\arguments{
\item{obj}{a zelig object with an estimated model}

\item{fitted}{a data frame with values fitted by \code{setx}. Note
created internally by \code{\link{extract_setx}} and
  \code{\link{extract_setrange}}}
}
\description{
Return individual factor coefficient fitted values to single factor variable
}
\author{
Christopher Gandrud
}
\keyword{internal}
