% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factorize}
\alias{factorize}
\title{Convert \code{as.factor}s called inside a \code{zelig} call}
\usage{
factorize(formula, data, check, f_out, d_out)
}
\arguments{
\item{formula}{model formulae}

\item{data}{data frame used in \code{formula}}

\item{check}{logical whether to just check if a formula contains internally
called factors and return \code{TRUE} or \code{FALSE}}

\item{f_out}{logical whether to return the converted formula}

\item{d_out}{logical whether to return the converted data frame. Note:
\code{f_out} must be missing}
}
\description{
Convert \code{as.factor}s called inside a \code{zelig} call
}
\author{
Christopher Gandrud
}
