% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{fitted,Zelig-method}
\alias{fitted,Zelig-method}
\title{Method for extracting estimated fitted values from Zelig objects}
\usage{
\S4method{fitted}{Zelig}(object, ...)
}
\arguments{
\item{object}{An Object of Class Zelig}

\item{...}{Additional parameters to be passed to fitted}
}
\description{
Method for extracting estimated fitted values from Zelig objects
}

