\name{friendship}

\alias{friendship}

\title{Simulated Example of Schoolchildren Friendship Network}

\description{
  This data set contains five sociomatrices of simulated data on friendship ties among schoolchildren.}

\usage{data(friendship)}

\format{
Each variable in the dataset is a 15 by 15 matrix representing some form of social network tie held by the fictitious children. The matrices are labeled "friends", "advice", "prestige", "authority", and "perpower".
}
\source{fictitious}

\keyword{datasets}


