% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_all_qi}
\alias{get_all_qi}
\title{Extract quantities of interest from a Zelig model}
\usage{
get_all_qi(z.out, nsim)
}
\arguments{
\item{z.out}{the output object from zelig}

\item{nsim}{the number of simulations used}
}
\value{
A list containing data.frames with columns stroring covariate values
  and quantities of interest. Specifically,
  \item{at} A data.frame with
  quantities of interest at specific covariate value and possibly
  \item{fd_at}{A data.frame with differences in quantities of interest at a
  set of covariate values}
}
\description{
Extract quantities of interest from a Zelig model
}

