% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_pvalue}
\alias{get_pvalue}
\title{Extract p-values from a Zelig estimated model}
\usage{
get_pvalue(object)
}
\arguments{
\item{object}{an object of class Zelig}
}
\description{
Extract p-values from a Zelig estimated model
}
\author{
Christopher Gandrud
}
