% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_se}
\alias{get_se}
\title{Extract standard errors from a Zelig estimated model}
\usage{
get_se(object)
}
\arguments{
\item{object}{an object of class Zelig}
}
\description{
Extract standard errors from a Zelig estimated model
}
\author{
Christopher Gandrud
}
