\name{help.zelig}

\alias{help.zelig}

\title{HTML Help for Zelig Commands and Models}

\description{
  The \code{help.zelig} command launches html help for Zelig commands
  and supported models.  The full manual is available online at
  \url{http://gking.harvard.edu/zelig}.
  }

\usage{
help.zelig(object)
}

\arguments{
  \item{object}{a character string representing a Zelig command or model. 
    \code{help.zelig("command")} will take you to an index of Zelig
    commands and \code{help.zelig("model")} will take you to a list of
    models. }
  }

\seealso{The complete document is available online at
  \url{http://gking.harvard.edu/zelig}.  
}

\references{Gary King, Michael Tomz, and Jason Wittenberg (2000),
``Making the Most of Statistical Analyses: Improving Interpretation and
  Presentation,'' \emph{American Journal of Political Science}, vol. 44,
  pp.347--355.
}

\author{
  Kosuke Imai <\email{kimai@princeton.edu}>; Gary King
  <\email{king@harvard.edu}>; Olivia Lau<\email{olau@fas.harvard.edu}>
}

\keyword{documentation}








