\name{hoff}

\alias{hoff}

\title{Social Security Expenditure Data}

\description{
  This data set contains annual social security expenditure (as percent
  of budget lagged by two years), the
  relative frequency of mentions social justice received in the party's
  platform in each year, and whether the president is Republican or
  Democrat.  
}

\usage{data(hoff)}

\format{A table containing 5 variables ("year", "L2SocSec", "Just503D", "Just503R", "RGovDumy") and 36 observations.}

\source{ICPSR (replication dataset s1109)}

\references{
  Gary King and Michael Laver. ``On Party Platforms, Mandates, and
  Government Spending,'' \emph{American Political Science Review},
  Vol. 87, No. 3 (September, 1993): pp. 744-750.
}

\keyword{datasets}
