\name{house}

\alias{house}

\title{Elections to the United States House of Representatives}

\description{
  This data set contains data on the 1984 and 1986 elections for the US
House of Representatives. }

\usage{data(house)}

\format{
  A table containing 12 variables and 320 observations.  Districts that 
were uncontested were omitted from the final data.  
}

\source{ICPSR}

\references{
  Andrew Gelman and Gary King.  ``Estimating Incumbency Advantage Without 
      Bias.'' American Journal of Political Science 34 (1990), 1142-1164.

  King, Gary.  ICPSR Publication Related Archive, 6311.
  
}

\keyword{datasets}


