\name{immigration}

\alias{immigration}
\alias{immi1}
\alias{immi2}
\alias{immi3}
\alias{immi4}
\alias{immi5}

\title{Individual Preferences Over Immigration Policy}

\description{These five datasets are part of a larger set of 10 multiply
  imputed data sets describing individual preferences toward immigration
  policy.  Imputation was performed via Amelia.  
}

\format{
  Each multiply-inputed data set consists of a table with 7 variables
  ("ipip", "wage1992", "prtyid",
  "ideol", "gender") and 2,485 observations.  For variable descriptions,
  please refer to Scheve and
  Slaugher, 2001.  
}

\source{National Election Survey}

\references{
 Scheve, Kenneth and Matthew Slaughter (2001). ``Labor Market Competition
 and Individual Preferences Over Immigration Policy,'' \emph{The Review of
 Economics and Statistics}, vol. 83, no. 1, pp. 133-145.  }

\keyword{datasets}


