% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_length_not_1}
\alias{is_length_not_1}
\title{Check if an object has a length greater than 1}
\usage{
is_length_not_1(x, msg = "Length is 1.", fail = TRUE)
}
\arguments{
\item{x}{an object}

\item{msg}{character string with the error message to return if
\code{fail = TRUE}.}

\item{fail}{logical whether to return an error if length is not greater than
1.}
}
\description{
Check if an object has a length greater than 1
}

