% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_sims_present}
\alias{is_sims_present}
\title{Check if any simulations are present in sim.out}
\usage{
is_sims_present(x, fail = TRUE)
}
\arguments{
\item{x}{a sim.out method}

\item{fail}{logical whether to return an error if no simulations are present.}
}
\description{
Check if any simulations are present in sim.out
}

