% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_simspresent}
\alias{is_simspresent}
\title{Check if simulations are present in sim.out}
\usage{
is_simspresent(x, fail = TRUE)
}
\arguments{
\item{x}{a sim.out method}

\item{fail}{logical whether to return an error if no simulations are present.}
}
\description{
Check if simulations are present in sim.out
}

