% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_simsrange1}
\alias{is_simsrange1}
\title{Check if simulations for a range1 of fitted values are present in sim.out}
\usage{
is_simsrange1(x, fail = TRUE)
}
\arguments{
\item{x}{a sim.out method}

\item{fail}{logical whether to return an error if simulation range is not
present.}
}
\description{
Check if simulations for a range1 of fitted values are present in sim.out
}
