% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_simsx}
\alias{is_simsx}
\title{Check if simulations for individual values are present in sim.out}
\usage{
is_simsx(x, fail = TRUE)
}
\arguments{
\item{x}{a sim.out method}

\item{fail}{logical whether to return an error if simulation range is not
present.}
}
\description{
Check if simulations for individual values are present in sim.out
}
