% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_simsx1}
\alias{is_simsx1}
\title{Check if simulations for individual values for first diferences are present
  in sim.out}
\usage{
is_simsx1(x, fail = TRUE)
}
\arguments{
\item{x}{a sim.out method}

\item{fail}{logical whether to return an error if simulation range is not
present.}
}
\description{
Check if simulations for individual values for first diferences are present
  in sim.out
}
