% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_timeseries}
\alias{is_timeseries}
\title{Check if a zelig object contains a time series model}
\usage{
is_timeseries(x, msg = "Not a timeseries object.", fail = FALSE)
}
\arguments{
\item{x}{a zelig object}

\item{msg}{character string with the error message to return if
\code{fail = TRUE}.}

\item{fail}{logical whether to return an error if \code{x} is not a timeseries.}
}
\description{
Check if a zelig object contains a time series model
}
