% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_uninitializedField}
\alias{is_uninitializedField}
\title{Check if uninitializedField}
\usage{
is_uninitializedField(
  x,
  msg = "Zelig model has not been estimated.",
  fail = TRUE
)
}
\arguments{
\item{x}{a zelig.out method}

\item{msg}{character string with the error message to return if
\code{fail = TRUE}.}

\item{fail}{logical whether to return an error if x uninitialzed.}
}
\description{
Check if uninitializedField
}
