% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_varying}
\alias{is_varying}
\title{Check if the values in a vector vary}
\usage{
is_varying(x, msg = "Vector does not vary.", fail = TRUE)
}
\arguments{
\item{x}{a vector}

\item{msg}{character string with the error message to return if
\code{fail = TRUE}.}

\item{fail}{logical whether to return an error if \code{x} does not vary.}
}
\description{
Check if the values in a vector vary
}
