% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_zelig}
\alias{is_zelig}
\title{Check if is a zelig object}
\usage{
is_zelig(x, fail = TRUE)
}
\arguments{
\item{x}{an object}

\item{fail}{logical whether to return an error if x is not a Zelig object.}
}
\description{
Check if is a zelig object
}

