% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_zeligei}
\alias{is_zeligei}
\title{Check if an object was created with ZeligEI}
\usage{
is_zeligei(x, msg = "Function is not relevant for ZeligEI objects.",
  fail = TRUE)
}
\arguments{
\item{x}{a zelig object}

\item{msg}{character string with the error message to return if
\code{fail = TRUE}.}

\item{fail}{logical whether to return an error if \code{x} is not a timeseries.}
}
\description{
Check if an object was created with ZeligEI
}
