\name{klein}

\alias{klein}

\title{Simulation Data for model Two-Stage Least Square (twosls) that corresponds to method 2SLS of systemfit}

\description{
  Dataframe contains annual observations of US economy from 1920 to
  1940. The columns are, Year, C=Consumption, P=Corporate profits,
  P1=Previous year corporate profit,Wtot=Total wage, Wp=Private wage
  bill, Wg=Government wage bill,I=Investment,
  K1=Previous year capital stock,X=GNP,G=Government spending, T=Taxes,
  X1=Previous year GNP, Tm=Year-1931.
  
}

\usage{data(klein)}

\format{
  A table containing 14 variables ("year", "C", "P", "P1","Wtot", "Wp",
  "Wg", "I", "K1","X", "G", "T", "X1", "Tm") and 21 observations.  
}
\source{http://pages.stern.nyu.edu/~wgreene/Text/econometricanalysis.htm}

\keyword{datasets}
