\name{kmenta}

\alias{kmenta}

\title{Simulation Data for model Three-Stage Least Square (threesls) that corresponds to method 3SLS of systemfit}

\description{
  Dataframe contains 20 annual observations of a supply/demand model
  with 5 variables. Columns are q=Food consumption per capita,
  p=Ratio of food price to general consumer prices, 
  d=Disposable income in contstant dollars,
  f=Ratio of preceding year's prices received by farmers to general consumer prices,
  a=Time index.
   
}

\usage{data(kmenta)}

\format{
  A table containing 5 variables ("q", "p", "d", "f","a")
  and 20 observations.  
}

\keyword{datasets}
