% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Median}
\alias{Median}
\title{Compute the Statistical Median of a Vector}
\usage{
Median(x, na.rm = NULL)
}
\arguments{
\item{x}{a vector of numeric or ordered values}

\item{na.rm}{ignored}
}
\value{
the median of the vector
}
\description{
Compute the Statistical Median of a Vector
}
\author{
Matt Owen \email{mowen@iq.harvard.edu}
}

