\name{mexico}

\alias{mexico}

\title{Voting Data from the 1988 Mexican Presidental Election}

\description{
  This dataset contains voting data for the 1988 Mexican presidential
  election.  
}

\usage{data(mexico)}

\format{A table containing 33 variables and 1,359 observations.}

\source{ICPSR}

\references{
  King, Gary, Michael Tomz and Jason Wittenberg (2000).
  ``Making the Most of Statistical Analyses: Improving Interpretation and 
  Presentation,'' \emph{American Journal of Political Science}, vol. 44,
  pp. 341-355.

  King, Tomz and Wittenberg.  ICPSR Publication Related Archive, 1255.  
}

\keyword{datasets}

