\name{mid}

\alias{mid}

\title{Militarized Interstate Disputes}

\description{
  A small sample from the militarized interstate disputes database,
  available at \url{http://pss.la.psu.edu/MID_DATA.HTM}.  
}

\usage{data(mid)}

\format{
  A table containing 6 variables ("conflict", "major", "contig", 
  "power", "maxdem", "mindem", and "years") and 3,126 observations.  For
  full variable descriptions, please see King and Zeng, 2001.  
}

\source{Militarized Interstate Disputes database}

\references{
  King, Gary, and Lanche Zeng (2001).  ``Explaining Rare Events in
  International Relations,'' \emph{International Organization}, vol. 55,
  no. 3, pp. 693-715.  

  Jones, Daniel M., Stuart A. Bremer and David Singer (1996).  ``Militarized
  Interstate Disputes, 1816-1992: Rationale, Coding Rules, and Empirical
  Patterns,'' \emph{Conflict Management and Peace Science}, vol. 15,
  no. 2, pp. 163-213.  
}

\keyword{datasets}


