% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\alias{mode}
\title{Compute the Statistical Mode of a Vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector of numeric, factor, or ordered values}
}
\value{
the statistical mode of the vector. If more than one mode exists,
 the last one in the factor order is arbitrarily chosen (by design)
}
\description{
Compute the Statistical Mode of a Vector
}
\author{
Christopher Gandrud and Matt Owen
}
