% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{model_lookup_df}
\alias{model_lookup_df}
\title{Instructions for how to convert non-Zelig fitted model objects to Zelig.
Used in to_zelig}
\format{An object of class \code{data.frame} with 9 rows and 4 columns.}
\usage{
model_lookup_df
}
\description{
Instructions for how to convert non-Zelig fitted model objects to Zelig.
Used in to_zelig
}
\keyword{datasets}
