% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_matcher}
\alias{model_matcher}
\title{Convenience function to find zelig class corresponding to a model fit.}
\usage{
model_matcher(obj)
}
\arguments{
\item{obj}{A model fit object.}
}
\description{
Convenience function to find zelig class corresponding to a model fit.
}
\keyword{internal}
