% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\name{names,Zelig-method}
\alias{names,Zelig-method}
\title{Names method for Zelig objects}
\usage{
\S4method{names}{Zelig}(x)
}
\arguments{
\item{x}{An Object of Class Zelig}
}
\description{
Names method for Zelig objects
}
