% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_pull}
\alias{p_pull}
\title{Extract p-values from a fitted model object}
\usage{
p_pull(x)
}
\arguments{
\item{x}{a fitted Zelig object}
}
\description{
Extract p-values from a fitted model object
}
\keyword{internal}
