% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{plot,Zelig,ANY-method}
\alias{plot,Zelig,ANY-method}
\title{Plot method for Zelig objects}
\usage{
\S4method{plot}{Zelig,ANY}(x, y, ...)
}
\arguments{
\item{x}{An Object of Class Zelig}

\item{y}{unused}

\item{...}{Additional parameters to be passed to plot}
}
\description{
Plot method for Zelig objects
}

