% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ci}
\alias{plot.ci}
\title{Method for plotting qi simulations across a range within a variable, with confidence intervals}
\usage{
\method{plot}{ci}(obj, qi="ev", var=NULL, ..., legcol="gray20", col=NULL, leg=1, legpos=NULL)
}
\arguments{
\item{qi}{a character-string specifying the quantity of interest to plot}

\item{var}{The variable to be used on the x-axis. Default is the variable
across all the chosen values with smallest nonzero variance}

\item{...}{Parameters to be passed to the `truehist' function which is
implicitly called for numeric simulations}

\item{main}{a character-string specifying the main heading of the plot}

\item{sub}{a character-string specifying the sub heading of the plot}

\item{xlab}{a character-string specifying the label for the x-axis}

\item{ylab}{a character-string specifying the label for the y-axis}

\item{legcol}{``legend color'', an valid color used for plotting the line
colors in the legend}

\item{col}{a valid vector of colors of at least length 3 to use to color the
confidence intervals}

\item{leg}{``legend position'', an integer from 1 to 4, specifying the
position of the legend. 1 to 4 correspond to ``SE'', ``SW'', ``NW'', and
``NE'' respectively}

\item{legpos}{``legend type'', exact coordinates and sizes for legend.
Overrides argment ``leg.type''}

\item{x}{A reference class zelig5 object}
}
\value{
the current graphical parameters. This is subject to change in future
implementations of Zelig
}
\description{
Method for plotting qi simulations across a range within a variable, with confidence intervals
}
\author{
James Honaker
}

