% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.qi}
\alias{plot.qi}
\title{Default Plot Design For Zelig Model QI's}
\usage{
plot.qi(obj, ...)
}
\description{
Default Plot Design For Zelig Model QI's
}
\author{
James Honaker with panel layouts from Matt Owen
}

