% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{plot.qidist}
\alias{plot.qidist}
\title{Plot distributions of quantities of interest.}
\usage{
plot.qidist(qi_dist)
}
\arguments{
\item{qi_dist}{A list containing distributions of quantities of interest.}
}
\value{
A ggplot object.
}
\description{
Plot distributions of quantities of interest.
}

