% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{predict,Zelig-method}
\alias{predict,Zelig-method}
\title{Method for getting predicted values from Zelig objects}
\usage{
\S4method{predict}{Zelig}(object, ...)
}
\arguments{
\item{object}{An Object of Class Zelig}

\item{...}{Additional parameters to be passed to predict}
}
\description{
Method for getting predicted values from Zelig objects
}

