% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qi.plot}
\alias{qi.plot}
\title{Default Plot Design For Zelig Model QI's}
\usage{
qi.plot(obj, ...)
}
\arguments{
\item{obj}{A reference class zelig5 object}

\item{...}{Parameters to be passed to the `truehist' function which is 
implicitly called for numeric simulations}
}
\description{
Default Plot Design For Zelig Model QI's
}
\author{
James Honaker with panel layouts from Matt Owen
}

