% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qi}
\alias{qi}
\title{Extract quantities of interest from a Zelig simulation}
\usage{
qi(object, qi = "ev", xvalue = "x", subset = NULL)
}
\arguments{
\item{object}{An Object of Class Zelig}

\item{qi}{character string with the name of quantity of interest desired:
\code{"ev"} for expected values, \code{"pv"} for predicted values or
\code{"fd"} for first differences.}

\item{xvalue}{chracter string stating which of the set values of `x`` should
be used for getting the quantity of interest.}

\item{subset}{subset for multiply imputed data (only relevant if multiply
imputed data is supplied in the original call.)}
}
\description{
Extract quantities of interest from a Zelig simulation
}
\author{
Christopher Gandrud
}
