% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{qi_dist}
\alias{qi_dist}
\alias{qi_dist.lm}
\title{Simulated distributions of quantities of interest.}
\usage{
qi_dist(model, nsim = 1000, ...)

\method{qi_dist}{lm}(model, at = list(), fd_at = list(x1 = NULL, x2 = NULL),
  nsim = 1000)
}
\arguments{
\item{model}{A model fit object}

\item{nsim}{the number of simulations used}

\item{...}{arguments passed to methods}

\item{at}{A named list of covariate values.}

\item{fd_at}{A list of length 2. Each element must itself be a named list of covariate values.}

\item{model}{A model fit object}

\item{nsim}{The number of simulations used}
}
\value{
A list containing data.frames with columns stroring covariate values
  and quantities of interest. Specifically,
  \item{at} A data.frame with
  quantities of interest at specific covariate value and possibly
  \item{fd_at}{A data.frame with differences in quantities of interest at a
  set of covariate values}

A list containing data.frames with columns stroring covariate values
  and quantities of interest. Specifically,
  \item{at} {A data.frame with simulated
  quantities of interest at specific covariate value}
  and possibly
  \item{fd_at}{A data.frame with differences in quantities of interest at a
  set of covariate values}
}
\description{
This is a generic function. Currently the only useful method is for lm
models.

Simulated distributions of quantities of interest from lm and glm models.
}

