% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce}
\alias{reduce}
\title{Describe Here}
\usage{
reduce(dataset, s, formula, data)
}
\arguments{
\item{dataset}{Describe Here}

\item{s}{Describe Here}

\item{formula}{Describe Here}

\item{data}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\author{
Christine Choirat
}
\keyword{internal}

