% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\name{residuals,Zelig-method}
\alias{residuals,Zelig-method}
\title{Method for extracting residuals from Zelig objects}
\usage{
\S4method{residuals}{Zelig}(object)
}
\arguments{
\item{object}{An Object of Class Zelig}
}
\description{
Method for extracting residuals from Zelig objects
}
