% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_intercept}
\alias{rm_intercept}
\title{Drop intercept columns from a data frame of fitted values}
\usage{
rm_intercept(x)
}
\arguments{
\item{x}{a data frame}
}
\description{
Drop intercept columns from a data frame of fitted values
}
\keyword{internal}
