\name{sanction}

\alias{sanction}

\title{Multilateral Economic Sanctions}

\description{
  Data on bilateral sanctions behavior for selected years during the
  general period 1939-1983.  This data contains errors that have since
  been corrected.  Please contact Lisa Martin before using this data for
  publication.   
}

\usage{data(sanction)}

\format{A table containing 8 variables ("mil", "coop", "target",
  "import", "export", "cost", "num", and "ncost") and 78 observations.
  For full variable description, see Martin, 1992. }

\source{Martin, 1992}

\references{
  Martin, Lisa (1992).  \emph{Coercive Cooperation: Explaining Multilateral
    Economic Sanctions}, Princeton: Princeton University Press.
}

\keyword{datasets}
