% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se_pull}
\alias{se_pull}
\title{Extract standard errors from a fitted model object}
\usage{
se_pull(x)
}
\arguments{
\item{x}{a fitted Zelig object}
}
\description{
Extract standard errors from a fitted model object
}
\keyword{internal}
