\name{seatshare}

\alias{seatshare}

\title{Leftist Party Seat Share in 11 OECD Countries}

\description{
  This data set contains time-series data of the seat shares in the lower legislative house of left leaning parties over time, as well as the level of unemployment.  Data follows the style used in Hibbs (1977).}

\usage{data(seatshare)}

\format{A table containing N variables ("country","year","unemp","leftseat") and 384 observations split across 11 countries.}

\source{OECD data and Mackie and Rose (1991), extended to further years.}

\references{
	Douglas A. Hibbs. (1977).  \emph{Political Parties and Macroeconomic Policy}. American Political Science Review 71(4):1467-1487.

	Thomas T. Mackie and Richard Rose.  (1991).  \emph{The International Almanac of Electoral History}  Macmillan: London.
}

\keyword{datasets}


