% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{setfactor}
\alias{setfactor}
\title{Describe Here}
\usage{
setfactor(fv, v)
}
\arguments{
\item{fv}{Describe Here}

\item{v}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\keyword{internal}

