% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{setval}
\alias{setval}
\title{Describe Here}
\usage{
setval(val, newval)
}
\arguments{
\item{val}{a vector of values}

\item{newval}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\keyword{internal}

