% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\name{simacf}
\alias{simacf}
\title{Construct Autocorrelation Function from Zelig object and simulated parameters}
\usage{
simacf(coef, order, params, alpha = 0.5)
}
\description{
Construct Autocorrelation Function from Zelig object and simulated parameters
}
\keyword{internal}
