\name{sna.ex}

\alias{sna.ex}

\title{Simulated Example of Social Network Data}

\description{
  This data set contains five sociomatrices of simulated data social network data.}

\usage{data(sna.ex)}

\format{
Each variable in the dataset is a 25 by 25 matrix of simulated social network data. The matrices are labeled "Var1", "Var2", "Var3", "Var4", and "Var5".
}

\source{fictitious}

\keyword{datasets}


