% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{statlevel}
\alias{statlevel}
\title{Describe Here}
\usage{
statlevel(qi, num)
}
\arguments{
\item{qi}{Describe Here}

\item{num}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\author{
Christine Choirat
}
\keyword{internal}

