% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{statmat}
\alias{statmat}
\title{Describe Here}
\usage{
statmat(qi)
}
\arguments{
\item{qi}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\author{
Christine Choirat
}
\keyword{internal}

