% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_package_name}
\alias{strip_package_name}
\title{Remove package names from fitted model object calls.}
\usage{
strip_package_name(x)
}
\arguments{
\item{x}{a fitted model object result}
}
\description{
Enables \code{\link{from_zelig_model}} output to work with stargazer.
}
\keyword{internal}
