% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{summary,Zelig-method}
\alias{summary,Zelig-method}
\title{Summary method for Zelig objects"}
\usage{
\S4method{summary}{Zelig}(object, ...)
}
\arguments{
\item{object}{An Object of Class Zelig}

\item{...}{Additional parameters to be passed to summary}
}
\description{
Summary method for Zelig objects"
}
