% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-timeseries.R
\name{summary.Arima}
\alias{summary.Arima}
\title{Summary of an object of class Arima}
\usage{
\method{summary}{Arima}(object, ...)
}
\arguments{
\item{object}{An object of class Arima}

\item{...}{Additional parameters}
}
\value{
The original object
}
\description{
Summary of an object of class Arima
}
