% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{summary.qidist}
\alias{summary.qidist}
\title{Summary statistic from distributions of quantities of interest.}
\usage{
summary.qidist(x, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{A list containing distributions of quantities of interest.}

\item{probs}{probabilities passed to quantile.}

\item{fd_at}{A list of length 2. Each element must itself be a named list of covariate values.}

\item{nsim}{The number of simulations used}
}
\value{
A list containing data.frames with columns stroring covariate values
  and summry statistic for quantities of interest. Specifically,
  \item{at} {A data.frame with simulated statistics for
  quantities of interest at specific covariate values}
  and possibly
  \item{fd_at}{A data.frame with summaries of differences in quantities of interest at a
  set of covariate values}
}
\description{
Summary statistic from distributions of quantities of interest.
}

