% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{table.levels}
\alias{table.levels}
\title{Describe Here}
\usage{
table.levels(x, levels, ...)
}
\arguments{
\item{x}{Describe Here}

\item{levels}{Describe Here}

\item{...}{Describe Here}
}
\value{
Describe Here
}
\description{
Describe Here
}
\keyword{internal}

