% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{to_zelig}
\alias{to_zelig}
\title{Coerce a non-Zelig fitted model object to a Zelig class object}
\usage{
to_zelig(obj)
}
\arguments{
\item{obj}{a fitted model object fitted using \code{lm}. MORE INTENDED}
}
\description{
Coerce a non-Zelig fitted model object to a Zelig class object
}
