\name{tobin}

\alias{tobin}

\title{Tobin's Tobit Data}

\description{
	Economists fit a parametric censored data model called the
     `tobit'. These data are from Tobin's original paper.
}

\usage{data(tobin)}

\format{
     A data frame with 20 observations on the following 3 variables.

     durable: Durable goods purchase

     age: Age in years

     quant: Liquidity ratio (x 1000)
}

\source{
   J. Tobin, Estimation of relationships for limited dependent
     variables, Econometrica, v26, 24-36, 1958.
   }

\keyword{datasets}
