% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformer}
\alias{transformer}
\title{Conduct variable transformations called inside a \code{zelig} call}
\usage{
transformer(formula, data, FUN = "log", check, f_out, d_out)
}
\arguments{
\item{formula}{model formulae}

\item{data}{data frame used in \code{formula}}

\item{FUN}{character string of the transformation function. Currently
supports \code{factor} and \code{log}.}

\item{check}{logical whether to just check if a formula contains an
internally called transformation and return \code{TRUE} or \code{FALSE}}

\item{f_out}{logical whether to return the converted formula}

\item{d_out}{logical whether to return the converted data frame. Note:
\code{f_out} must be missing}
}
\description{
Conduct variable transformations called inside a \code{zelig} call
}
\author{
Christopher Gandrud
}
\keyword{internal}
