\name{turnout}

\alias{turnout}

\title{Turnout Data Set from the National Election Survey}

\description{
  This data set contains individual-level turnout data. It pools several
  American National Election Surveys conducted during the 1992 presidential
  election year.  Only the first 2,000 observations (from a total of 15,837 
  observations) are included in the sample data.  
}

\usage{data(turnout)}

\format{A table containing 5 variables ("race", "age", "educate", 
"income", and "vote") and 2,000 observations.}

\source{National Election Survey}

\references{
  King, Gary, Michael Tomz, Jason Wittenberg (2000).
  ``Making the Most of Statistical Analyses: Improving Interpretation and 
  Presentation,'' \emph{American Journal of Political Science}, vol. 44,
  pp.341--355.
}

\keyword{datasets}
