% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\name{vcov,Zelig-method}
\alias{vcov,Zelig-method}
\title{Variance-covariance method for Zelig objects}
\usage{
\S4method{vcov}{Zelig}(object)
}
\arguments{
\item{object}{An Object of Class Zelig}
}
\description{
Variance-covariance method for Zelig objects
}
