\name{Weimar}
\alias{Weimar}

\title{1932 Weimar election data}

\description{ This data set contains election results for 10 kreise (equivalent to precincts) from the 1932 Weimar (German) election.  
}

\usage{data(Weimar)}

\format{A table containing 11 variables and 10 observations.  The variables are {
{Nazi}{Number of votes for the Nazi party}
{Government}{Number of votes for the Government}
{Communists}{Number of votes for the Communist party}
{FarRight}{Number of votes for far right parties}
{Other}{Number of votes for other parties, and non-voters}
{shareunemployed}{Proportion unemployed}
{shareblue}{Proportion working class}
{sharewhite}{Proportion white-collar workers}
{sharedomestic}{Proportion domestic servants}
{shareprotestants}{Proportion Protestant} 
}}

\source{ICPSR}

%\references{
%}

\keyword{datasets}

