% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ar.R
\docType{class}
\name{Zelig-ar-class}
\alias{Zelig-ar-class}
\alias{zar}
\title{Time-Series Model with Autoregressive Disturbance}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-ar.html}
}

