% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\docType{class}
\name{Zelig-arima-class}
\alias{Zelig-arima-class}
\alias{zarima}
\title{Autoregressive and Moving-Average Models with Integration for Time-Series Data}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_arima.html}
}

