% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bayes.R
\docType{class}
\name{Zelig-bayes-class}
\alias{Zelig-bayes-class}
\alias{zbayes}
\title{Bayes Model object for inheritance across models in Zelig}
\description{
Bayes Model object for inheritance across models in Zelig
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig command estimates a variety of statistical models}
}}

