% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-binchoice.R
\docType{class}
\name{Zelig-binchoice-class}
\alias{Zelig-binchoice-class}
\alias{zbinchoice}
\title{Binary Choice object for inheritance across models in Zelig}
\description{
Binary Choice object for inheritance across models in Zelig
}

