% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-binchoice-gee.R
\docType{class}
\name{Zelig-binchoice-gee-class}
\alias{Zelig-binchoice-gee-class}
\alias{zbinchoicegee}
\title{Object for Binary Choice outcomes in Generalized Estimating Equations 
for inheritance across models in Zelig}
\description{
Object for Binary Choice outcomes in Generalized Estimating Equations 
for inheritance across models in Zelig
}

