% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-binchoice-survey.R
\docType{class}
\name{Zelig-binchoice-survey-class}
\alias{Zelig-binchoice-survey-class}
\alias{zbinchoicesurvey}
\title{Object for Binary Choice outcomes with Survey Weights
for inheritance across models in Zelig}
\description{
Object for Binary Choice outcomes with Survey Weights
for inheritance across models in Zelig
}

