% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-exp.R
\docType{class}
\name{Zelig-exp-class}
\alias{Zelig-exp-class}
\alias{zexp}
\title{Exponential Regression for Duration Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_exp.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE, refit = NULL)}}{The zelig function estimates a variety of statistical models}
}}

