% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-factor-bayes.R
\docType{class}
\name{Zelig-factor-bayes-class}
\alias{Zelig-factor-bayes-class}
\alias{zfactorbayes}
\title{Bayesian Factor Analysis}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_factorbayes.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE, refit = NULL)}}{The zelig function estimates a variety of statistical models}
}}

