% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gamma-gee.R
\docType{class}
\name{Zelig-gamma-gee-class}
\alias{Zelig-gamma-gee-class}
\alias{zgammagee}
\title{Generalized Estimating Equation for Gamma Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig-gammagee.html}
}

