% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-gamma-survey.R
\docType{class}
\name{Zelig-gamma-survey-class}
\alias{Zelig-gamma-survey-class}
\alias{zgammasurvey}
\title{Gamma Regression with Survey Weights}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-gammasurvey.html}
}

