% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gee.R
\docType{class}
\name{Zelig-gee-class}
\alias{Zelig-gee-class}
\alias{zgee}
\title{Generalized Estimating Equations Model object for inheritance across models in Zelig}
\description{
Generalized Estimating Equations Model object for inheritance across models in Zelig
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

