% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-glm.R
\docType{class}
\name{Zelig-glm-class}
\alias{Zelig-glm-class}
\alias{zglm}
\title{Generalized Linear Model object for inheritance across models in Zelig}
\description{
Generalized Linear Model object for inheritance across models in Zelig
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

