% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-ivreg.R
\docType{class}
\name{Zelig-ivreg-class}
\alias{Zelig-ivreg-class}
\alias{zivreg}
\title{Instrumental-Variable Regression}
\description{
Fit instrumental-variable regression by two-stage least squares. This is
equivalent to direct instrumental-variables estimation when the number of
instruments is equal to the number of predictors.
}
\details{
[FILL IN PARAMS]
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

