% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-logit-bayes.R
\docType{class}
\name{Zelig-logit-bayes-class}
\alias{Zelig-logit-bayes-class}
\alias{zlogitbayes}
\title{Bayesian Logit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_logitbayes.html}
}

