% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-logit-gee.R
\docType{class}
\name{Zelig-logit-gee-class}
\alias{Zelig-logit-gee-class}
\alias{zlogitgee}
\title{Generalized Estimating Equation for Logit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-logitgee.html}
}

