% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-logit-survey.R
\docType{class}
\name{Zelig-logit-survey-class}
\alias{Zelig-logit-survey-class}
\alias{zlogitsurvey}
\title{Logit Regression with Survey Weights}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig-logitsurvey.html}
}

