% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lognorm.R
\docType{class}
\name{Zelig-lognorm-class}
\alias{Zelig-lognorm-class}
\alias{zlognorm}
\title{Log-Normal Regression for Duration Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-lognorm.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

