% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-negbinom.R
\docType{class}
\name{Zelig-negbin-class}
\alias{Zelig-negbin-class}
\alias{znegbin}
\title{Negative Binomial Regression for Event Count Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-negbin.html}
}
\section{Methods}{

\describe{
\item{\code{test(z, data)}}{Monte Carlo Unit Test}

\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

