% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-normal.R
\docType{class}
\name{Zelig-normal-class}
\alias{Zelig-normal-class}
\alias{znormal}
\title{Normal Regression for Continuous Dependent Variables}
\arguments{
\item{formula}{a model fitting formula}
}
\description{
Normal Regression for Continuous Dependent Variables
}

\seealso{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_normal.html}
}
