% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-normal-gee.R
\docType{class}
\name{Zelig-normal-gee-class}
\alias{Zelig-normal-gee-class}
\alias{znormalgee}
\title{Generalized Estimating Equation for Normal Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-normalgee.html}
}

