% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-normal-survey.R
\docType{class}
\name{Zelig-normal-survey-class}
\alias{Zelig-normal-survey-class}
\alias{znormalsurvey}
\title{Normal Regression for Continuous Dependent Variables with Survey Weights}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-normalsurvey.html}
}

