% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-poisson-bayes.R
\docType{class}
\name{Zelig-poisson-bayes-class}
\alias{Zelig-poisson-bayes-class}
\alias{zpoissonbayes}
\title{Bayesian Poisson Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-poissonbayes.html}
}

