% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-poisson.R
\docType{class}
\name{Zelig-poisson-class}
\alias{Zelig-poisson-class}
\alias{zpoisson}
\title{Poisson Regression for Event Count Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-poisson.html}
}

