% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-poisson-gee.R
\docType{class}
\name{Zelig-poisson-gee-class}
\alias{Zelig-poisson-gee-class}
\alias{zpoissongee}
\title{Generalized Estimating Equation for Poisson Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-poissongee.html}
}

