% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-poisson-survey.R
\docType{class}
\name{Zelig-poisson-survey-class}
\alias{Zelig-poisson-survey-class}
\alias{zpoissonsurvey}
\title{Poisson Regression with Survey Weights}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-poissonsurvey.html}
}

