% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-probit.R
\docType{class}
\name{Zelig-probit-class}
\alias{Zelig-probit-class}
\alias{zprobit}
\title{Probit Regression for Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-probit.html}
}

