% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-probit-gee.R
\docType{class}
\name{Zelig-probit-gee-class}
\alias{Zelig-probit-gee-class}
\alias{zprobitgee}
\title{Generalized Estimating Equation for Probit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-probitgee.html}
}

