% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-probit-survey.R
\docType{class}
\name{Zelig-probit-survey-class}
\alias{Zelig-probit-survey-class}
\alias{zprobitsurvey}
\title{Probit Regression with Survey Weights}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_probitsurvey.html}
}

