% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-survey.R
\docType{class}
\name{Zelig-survey-class}
\alias{Zelig-survey-class}
\alias{zsurvey}
\title{Survey models in Zelig for weights for complex sampling designs}
\description{
Survey models in Zelig for weights for complex sampling designs
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE, refit = NULL)}}{The zelig function estimates a variety of statistical models}
}}

