% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-timeseries.R
\docType{class}
\name{Zelig-timeseries-class}
\alias{Zelig-timeseries-class}
\alias{ztimeseries}
\title{Time-series models in Zelig}
\description{
Time-series models in Zelig
}
\section{Methods}{

\describe{
\item{\code{packagename()}}{Automatically retrieve wrapped package name}

\item{\code{sim(num = NULL)}}{Generic Method for Computing and Organizing Simulated Quantities of Interest}

\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by, bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

