% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tobit-bayes.R
\docType{class}
\name{Zelig-tobit-bayes-class}
\alias{Zelig-tobit-bayes-class}
\alias{ztobitbayes}
\title{Bayesian Tobit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_tobitbayes.html}
}

