% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tobit.R
\docType{class}
\name{Zelig-tobit-class}
\alias{Zelig-tobit-class}
\alias{ztobit}
\title{Linear Regression for a Left-Censored Dependent Variable}
\description{
Linear Regression for a Left-Censored Dependent Variable
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

\seealso{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_tobit.html}
}
