% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zelig_mutate}
\alias{zelig_mutate}
\title{Zelig Copy of plyr::mutate to avoid namespace conflict with dplyr}
\source{
Hadley Wickham (2011). The Split-Apply-Combine Strategy for Data
Analysis. Journal of Statistical Software, 40(1), 1-29. URL
\url{http://www.jstatsoft.org/v40/i01/}.
}
\usage{
zelig_mutate(.data, ...)
}
\description{
Zelig Copy of plyr::mutate to avoid namespace conflict with dplyr
}
\keyword{internal}

