% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{zelig_qi_to_df}
\alias{zelig_qi_to_df}
\title{Extract simulated quantities of interest from a zelig object}
\source{
For a discussion of tidy data see
\url{https://vita.had.co.nz/papers/tidy-data.pdf}.
}
\usage{
zelig_qi_to_df(obj)
}
\arguments{
\item{obj}{a zelig object with simulated quantities of interest}
}
\description{
Extract simulated quantities of interest from a zelig object
}
\details{
Returns simulated quantities of interest in a tidy data formatted
\code{data.frame}. This can be useful for creating custom plots.

Each row contains a simulated value and each column contains:
\itemize{
\item \code{x} whether the simulations are from the base \code{x} \code{setx} or the
contrasting \code{x1} for finding first differences.
\item The fitted values specified in \code{setx} including a \code{by} column if
\code{by} was used in the \code{\link{zelig}} call.
\item \code{expected_value}
\item \code{expected_value}
}
}
\examples{
#### QIs without first difference or range, from covariates fitted at
## central tendencies
z4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")
z4 <- setx(z4)
z4 <- sim(z4)
zelig_qi_to_df(z4)

#### QIs for first differences
z4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")
z4.1 <- setx(z4, Petal.Length = 2)
z4.2 <- setx(z4, Petal.Length = 4.4)
z4 <- sim(z4, x = z4.1, x1 = z4.2)
zelig_qi_to_df(z4)

#### QIs for first differences, estimated by Species
z4 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")
z4.1 <- setx(z4, Petal.Length = 2)
z4.2 <- setx(z4, Petal.Length = 4.4)
z4 <- sim(z4, x = z4.1, x1 = z4.2)
zelig_qi_to_df(z4)

#### QIs for a range of fitted values
z4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")
z4 <- setx(z4, Petal.Length = 2:4)
z4 <- sim(z4)
zelig_qi_to_df(z4)

#### QIs for a range of fitted values, estimated by Species
z4 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
            model = "ls")
z4 <- setx(z4, Petal.Length = 2:4)
z4 <- sim(z4)
zelig_qi_to_df(z4)

#### QIs for two ranges of fitted values
z4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
            model = "ls")
z4.1 <- setx(z4, Petal.Length = 2:4, Species = 'setosa')
z4.2 <- setx(z4, Petal.Length = 2:4, Species = 'virginica')
z4 <- sim(z4, x = z4.1, x1 = z4.2)

zelig_qi_to_df(z4)

}
\author{
Christopher Gandrud
}
