% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{zeligACFplot}
\alias{zeligACFplot}
\title{Plot Autocorrelation Function from Zelig QI object}
\usage{
zeligACFplot(
  z,
  omitzero = FALSE,
  barcol = "black",
  epsilon = 0.1,
  col = NULL,
  main = "Autocorrelation Function",
  xlab = "Period",
  ylab = "Correlation of Present Shock with Future Outcomes",
  ylim = NULL,
  ...
)
}
\description{
Plot Autocorrelation Function from Zelig QI object
}
\keyword{internal}
