% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\name{zeligArimaWrapper}
\alias{zeligArimaWrapper}
\title{Estimation wrapper function for arima models, to easily fit with Zelig architecture}
\usage{
zeligArimaWrapper(formula, order = c(1, 0, 0), ..., include.mean = TRUE,
  data)
}
\description{
Estimation wrapper function for arima models, to easily fit with Zelig architecture
}
\keyword{internal}
