% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\name{zeligARMAbreakforecaster}
\alias{zeligARMAbreakforecaster}
\title{Construct Simulated Series with Internal Discontinuity in X}
\usage{
zeligARMAbreakforecaster(
  y.init = NULL,
  x,
  x1,
  simparam,
  order,
  sd,
  t1 = 5,
  t2 = 10
)
}
\description{
Construct Simulated Series with Internal Discontinuity in X
}
\keyword{internal}
