% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\name{zeligARMAlongrun}
\alias{zeligARMAlongrun}
\title{Calculate the Long Run Exquilibrium for Fixed X}
\usage{
zeligARMAlongrun(
  y.init = NULL,
  x,
  simparam,
  order,
  sd,
  tol = NULL,
  burnin = 20
)
}
\description{
Calculate the Long Run Exquilibrium for Fixed X
}
\keyword{internal}
