% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-arima.R
\name{zeligARMAnextstep}
\alias{zeligARMAnextstep}
\title{Construct Simulated Next Step in Dynamic Series}
\usage{
zeligARMAnextstep(
  yseries = NULL,
  xseries,
  wseries = NULL,
  beta,
  ar = NULL,
  i = NULL,
  ma = NULL,
  sd
)
}
\description{
Construct Simulated Next Step in Dynamic Series
}
\keyword{internal}
