% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zmodelMatcher.lm}
\alias{zmodelMatcher.lm}
\title{Match a model fit object to the corresponsding Zelig class.}
\usage{
\method{zmodelMatcher}{lm}(model)
}
\arguments{
\item{model}{A model fit object}
}
\value{
a zelig object capable of producing the model.
}
\description{
This is a generic function. At the moment the only useful method is for lm objects.
}

