% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zmodelMatcher}
\alias{zmodelMatcher}
\title{Match a model fit object to the corresponsding Zelig class.}
\usage{
zmodelMatcher(model, ...)
}
\arguments{
\item{model}{A model fit object}

\item{...}{Not currently used.}
}
\description{
This is a generic function. At the moment the only useful method is for lm objects.
}

