% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zmodelsAvailable}
\alias{zmodelsAvailable}
\title{List available zelig models.}
\usage{
zmodelsAvailable()
}
\value{
A named list. Each element contains information about a zelig class
}
\description{
List available zelig models.
}

