\name{approval}
\alias{approval}

\title{U.S. Presidential Approval Data}

\description{
  Monthy public opinion data for 2001-2006.  
}

\usage{data(approval)}

\format{
  A table containing 8 variables ("month", "year", "approve", 
  "disapprove", "unsure", "sept.oct.2001", "iraq.war", and "avg.price")
  and 65 observations.  }

\source{ICPSR}

\references{
  Stuff here
}

\keyword{datasets}




