% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster.formula}
\alias{cluster.formula}
\title{Generate Formulae that Consider Clustering}
\usage{
cluster.formula(formula, cluster)
}
\arguments{
\item{formula}{a formula object}

\item{cluster}{a vector}
}
\value{
a formula object describing clustering
}
\description{
This method is used internally by the "Zelig" Package to interpret
clustering in GEE models.
}

