\name{coalition}

\alias{coalition}

\title{Coalition Dissolution in Parliamentary Democracies}

\description{
  This data set contains survival data on government coalitions in
  parliamentary democracies (Belgium, Canada, Denmark, Finland, France,
  Iceland, Ireland, Israel, Italy, Netherlands, Norway, Portugal, Spain,
  Sweden, and the United Kingdom) for the period 1945-1987.  For
  parsimony, country indicator variables are omitted in the sample data.  
}

\usage{data(coalition)}

\format{
  A table containing 7 variables ("duration", "ciep12", "invest",
  "fract", "polar", "numst2", "crisis") and 314 observations.  For
  variable descriptions, please refer to King, Alt, Burns and Laver
  (1990).  
}

\source{ICPSR}

\references{
  King, Gary, James E. Alt, Nancy Elizabeth Burns and Michael Laver (1990).
  ``A Unified Model  of Cabinet Dissolution in Parliamentary
  Democracies,'' \emph{American Journal of Political Science}, vol. 34,
  no. 3, pp. 846-870.  
  
  Gary King, James E. Alt, Nancy Burns, and Michael Laver.  ICPSR
  Publication Related Archive, 1115.
}

\keyword{datasets}


