% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{coef,Zelig-method}
\alias{coef,Zelig-method}
\title{Method for extracting estimated coefficients from Zelig objects}
\usage{
\S4method{coef}{Zelig}(object)
}
\arguments{
\item{object}{An Object of Class Zelig}
}
\description{
Method for extracting estimated coefficients from Zelig objects
}
