% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mi}
\alias{mi}
\title{Enables backwards compatability for preparing non-amelia imputed data sets
for \code{zelig}.}
\usage{
mi(...)
}
\arguments{
\item{...}{a set of \code{data.frame}'s}
}
\value{
an \code{mi} object composed of a list of data frames.
}
\description{
See \code{\link{to_zelig_mi}}
}
