% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Compute the Statistical Mode of a Vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector of numeric, factor, or ordered values}
}
\value{
the statistical mode of the vector. If two modes exist, one is
  randomly selected (by design)
}
\description{
Compute the Statistical Mode of a Vector
}
\author{
Matt Owen \email{mowen@iq.harvard.edu}
}

