% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{or_summary}
\alias{or_summary}
\title{Find odds ratios for coefficients and standard errors
for glm.summary class objects}
\usage{
or_summary(obj, label_mod = "(OR)")
}
\description{
obj a \code{glm.summary} class object
}
