% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-relogit.R
\name{relogit}
\alias{relogit}
\title{Estimation function for rare events logit models}
\usage{
relogit(formula, data = sys.parent(), tau = NULL, bias.correct = TRUE,
  case.control = "prior", ...)
}
\description{
Estimation function for rare events logit models
}
\keyword{internal}

