% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-relogit.R
\name{relogit}
\alias{relogit}
\title{Internal function for Rare Events Logistic Regression for Dichotomous Dependent Variables}
\usage{
relogit(formula, data = sys.parent(), tau = NULL, bias.correct = TRUE,
  case.control = "prior", ...)
}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-relogit.html}
}
\keyword{internal}

