% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{setfactor}
\alias{setfactor}
\title{Set new value of a factor variable, checking for existing levels}
\usage{
setfactor(fv, v)
}
\arguments{
\item{fv}{factor variable}

\item{v}{value}
}
\value{
a factor variable with a value \code{val} and the same levels
}
\description{
Set new value of a factor variable, checking for existing levels
}
\keyword{internal}

