% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setval}
\alias{setval}
\title{Set new value of a variable as approrpriate to data type}
\usage{
setval(val, newval)
}
\arguments{
\item{val}{old value}

\item{newval}{new value}
}
\value{
a variable of the same type with a value \code{val}
}
\description{
Set new value of a variable as approrpriate to data type
}
\keyword{internal}

