% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat}
\alias{stat}
\title{Pass Quantities of Interest to Appropriate Summary Function}
\usage{
stat(qi, num)
}
\arguments{
\item{qi}{quantity of interest (e.g., estimated value or predicted value)}

\item{num}{number of simulations}
}
\value{
a formatted qi
}
\description{
Pass Quantities of Interest to Appropriate Summary Function
}
\author{
Christine Choirat
}
\keyword{internal}

