% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{stat}
\alias{stat}
\title{}
\usage{
stat(qi, num)
}
\arguments{
\item{qi}{quantity of interest (e.g., estimated value or predicted value)}

\item{num}{number of simulations}
}
\value{
a formatted qi
}
\description{

}
\author{
Christine Choirat
}
\keyword{internal}

