% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{statlevel}
\alias{statlevel}
\title{Summarize Quantities of Interest when Factors}
\usage{
statlevel(qi, num)
}
\arguments{
\item{qi}{quantity of interest in the discrete case}

\item{num}{number of simulations}
}
\value{
a formatted qi
}
\description{
Summarize Quantities of Interest when Factors
}
\author{
Christine Choirat
}
\keyword{internal}

