% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{statmat}
\alias{statmat}
\title{Create QI summary matrix}
\usage{
statmat(qi)
}
\arguments{
\item{qi}{quantity of interest in the discrete case}
}
\value{
a formatted qi
}
\description{
Create QI summary matrix
}
\author{
Christine Choirat
}
\keyword{internal}
