% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{statmat}
\alias{statmat}
\title{Summarize Quantities of Interest when Matricies}
\usage{
statmat(qi)
}
\arguments{
\item{qi}{quantity of interest in the discrete case}
}
\value{
a formatted qi
}
\description{
Summarize Quantities of Interest when Matricies
}
\author{
Christine Choirat
}
\keyword{internal}

