% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{to_zelig}
\alias{to_zelig}
\title{Coerce a non-Zelig fitted model object to a Zelig class object}
\usage{
to_zelig(obj, ...)
}
\arguments{
\item{obj}{a fitted model object fitted using \code{lm} and many using
\code{glm}. Note: more intended in future Zelig releases.}

\item{...}{other arguments passed to \code{zelig}.}
}
\description{
Coerce a non-Zelig fitted model object to a Zelig class object
}
\examples{
library(dplyr)
lm.out <- lm(Fertility ~ Education, data = swiss)

z.out <- to_zelig(lm.out)

# to_zelig called from within setx
setx(z.out) \%>\% sim() \%>\% plot()

if (require(mitools)) {
   data(africa, package = "Amelia")
   a.imp <- Amelia::amelia(x = africa, cs = "country", ts = "year", logs = "gdp_pc")$imputations
   mod.out <- with(imputationList(a.imp), lm(gdp_pc ~ trade + civlib))
   to_zelig(mod.out)
}

}
\author{
Christopher Gandrud and Ista Zhan
}
