% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-zelig.R
\docType{methods}
\name{vcov,Zelig-method}
\alias{vcov,Zelig-method}
\title{Variance-covariance method for Zelig objects}
\usage{
\S4method{vcov}{Zelig}(object, ...)
}
\arguments{
\item{object}{An Object of Class Zelig}

\item{...}{Additional parameters to be passed to vcov}
}
\description{
Variance-covariance method for Zelig objects
}
