% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Weimar}
\alias{Weimar}
\title{1932 Weimar election data}
\format{
A table containing 11 variables and 10 observations. The variables are
\describe{
\item{Nazi}{Number of votes for the Nazi party}
\item{Government}{Number of votes for the Government}
\item{Communists}{Number of votes for the Communist party}
\item{FarRight}{Number of votes for far right parties}
\item{Other}{Number of votes for other parties, and non-voters}
\item{shareunemployed}{Proportion unemployed}
\item{shareblue}{Proportion working class}
\item{sharewhite}{Proportion white-collar workers}
\item{sharedomestic}{Proportion domestic servants}
\item{shareprotestants}{Proportion Protestant}
}
}
\source{
ICPSR
}
\description{
1932 Weimar election data
}
\keyword{datasets}
