% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-factor-bayes.R
\docType{class}
\name{Zelig-factor-bayes-class}
\alias{Zelig-factor-bayes-class}
\alias{zfactorbayes}
\title{Bayesian Factor Analysis}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-factorbayes.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models.}
}}

