% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gamma.R
\docType{class}
\name{Zelig-gamma-class}
\alias{Zelig-gamma-class}
\alias{zgamma}
\title{Gamma Regression for Continuous, Positive Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig-gamma.html}
}

