% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-logit-bayes.R
\docType{class}
\name{Zelig-logit-bayes-class}
\alias{Zelig-logit-bayes-class}
\alias{zlogitbayes}
\title{Bayesian Logit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-logitbayes.html}
}

