% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-logit.R
\docType{class}
\name{Zelig-logit-class}
\alias{Zelig-logit-class}
\alias{zlogit}
\title{Logistic Regression for Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_logit.html}
}
\section{Methods}{

\describe{
\item{\code{show(signif.stars = FALSE, subset = NULL, bagging = FALSE)}}{Display a Zelig object}
}}

