% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-logit.R
\docType{class}
\name{Zelig-logit-class}
\alias{Zelig-logit-class}
\alias{zlogit}
\title{Logistic Regression for Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-logit.html}
}

