% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-ls.R
\docType{class}
\name{Zelig-ls-class}
\alias{Zelig-ls-class}
\alias{zls}
\title{Least Squares Regression for Continuous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-ls.html}
}
\section{Methods}{

\describe{
\item{\code{test(z, data)}}{Monte Carlo Unit Test}

\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

