% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mlogit-bayes.R
\docType{class}
\name{Zelig-mlogit-bayes-class}
\alias{Zelig-mlogit-bayes-class}
\alias{zmlogitbayes}
\title{Bayesian Multinomial Logistic Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig-mlogitbayes.html}
}

