% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-negbinom.R
\docType{class}
\name{Zelig-negbin-class}
\alias{Zelig-negbin-class}
\alias{znegbin}
\title{Negative Binomial Regression for Event Count Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_negbin.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

