% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-normal-bayes.R
\docType{class}
\name{Zelig-normal-bayes-class}
\alias{Zelig-normal-bayes-class}
\alias{znormalbayes}
\title{Bayesian Normal Linear Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig-normalbayes.html}
}

