% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-normal.R
\docType{class}
\name{Zelig-normal-class}
\alias{Zelig-normal-class}
\alias{znormal}
\title{Normal Regression for Continuous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-normal.html}
}

