% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-oprobit-bayes.R
\docType{class}
\name{Zelig-oprobit-bayes-class}
\alias{Zelig-oprobit-bayes-class}
\alias{zoprobitbayes}
\title{Bayesian Ordered Probit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_oprobitbayes.html}
}

