% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-probit-bayes.R
\docType{class}
\name{Zelig-probit-bayes-class}
\alias{Zelig-probit-bayes-class}
\alias{zprobitbayes}
\title{Bayesian Probit Regression}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-probitbayes.html}
}

