% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-quantile.R
\docType{class}
\name{Zelig-quantile-class}
\alias{Zelig-quantile-class}
\alias{zquantile}
\title{Quantile Regression for Continuous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-quantile.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

