% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-relogit.R
\docType{class}
\name{Zelig-relogit-class}
\alias{Zelig-relogit-class}
\alias{zrelogit}
\title{Rare Events Logistic Regression for Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-relogit.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models.}
}}

