% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-relogit.R
\docType{class}
\name{Zelig-relogit-class}
\alias{Zelig-relogit-class}
\alias{zrelogit}
\title{Rare Events Logistic Regression for Dichotomous Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-relogit.html}
}
\section{Methods}{

\describe{
\item{\code{modcall_formula_transformer()}}{Transform model call formula.}

\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

