% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model-tobit.R
\docType{class}
\name{Zelig-tobit-class}
\alias{Zelig-tobit-class}
\alias{ztobit}
\title{Linear Regression for a Left-Censored Dependent Variable}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-tobit.html}
}
\section{Methods}{

\describe{
\item{\code{test(z, data)}}{Monte Carlo Unit Test}

\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models.}
}}

