% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-weibull.R
\docType{class}
\name{Zelig-weibull-class}
\alias{Zelig-weibull-class}
\alias{zweibull}
\title{Weibull Regression for Duration Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/articles/zelig_weibull.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, model = NULL, ..., weights = NULL, by,
  bootstrap = FALSE)}}{The zelig function estimates a variety of statistical models}
}}

