% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model-weibull.R
\docType{class}
\name{Zelig-weibull-class}
\alias{Zelig-weibull-class}
\alias{zweibull}
\title{Weibull Regression for Duration Dependent Variables}
\description{
Vignette: \url{http://docs.zeligproject.org/en/latest/zelig-weibull.html}
}
\section{Methods}{

\describe{
\item{\code{zelig(formula, data, ..., weights = NULL, by)}}{The zelig command estimates a variety of statistical models}
}}

