% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{zelig_qi_to_df}
\alias{zelig_qi_to_df}
\title{Extract simulated quantities of interest from a zelig object}
\source{
For a discussion of tidy data see
\url{https://vita.had.co.nz/papers/tidy-data.pdf}.
}
\usage{
zelig_qi_to_df(obj)
}
\arguments{
\item{obj}{a zelig object with simulated quantities of interest}
}
\description{
Extract simulated quantities of interest from a zelig object
}
\details{
Returns simulated quantities of interest in a tidy data formatted
\code{data.frame}. This can be useful for creating custom plots.

Each row contains a simulated value and each column contains:
- The fitted values specified in \code{setx}
- \code{ev}: expected values
- \code{pv}: predicted values
}
\examples{
    z4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
                   model = "ls")
    z4 <- setx(z4)
    z4 <- sim(z4)
    zelig_qi_to_df(z4)

}
\author{
Christopher Gandrud
}
