% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeligPlyrMutate}
\alias{zeligPlyrMutate}
\title{Zelig Copy of plyr::mutate to avoid namespace conflict with dplyr}
\usage{
zeligPlyrMutate(.data, ...)
}
\description{
Zelig Copy of plyr::mutate to avoid namespace conflict with dplyr
}
\keyword{internal}

